requirements.txt
aiomysql==0.3.2
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astor==0.8.1
async-timeout==4.0.3
bce-python-sdk==0.9.52
Brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
hyperframe==6.1.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-text-splitters==0.3.8
langsmith==0.4.43
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.18
modelscope==1.31.0
networkx==3.4.2
numpy==2.2.6
openai==2.8.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
orjson==3.11.4
packaging==24.2
paddleocr==3.3.2
paddlepaddle==3.0.0
paddlex==3.3.9
pandas==2.3.3
pillow==12.0.0
pip==24.3.1
prettytable==3.17.0
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.12.0
Pygments==2.19.2
PyMySQL==1.1.2
pypdfium2==5.0.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
setuptools==44.1.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.2
tenacity==9.1.2
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
zstandard==0.25.0
